Fix warning from MSVC 64-bit.

+++ test/unit/testSplitVector.cxx
@@ -274,8 +274,8 @@ TEST_CASE("SplitVector") {
 		for (int i=0; i<testLength; i++)
 			sv.SetValueAt(i, i+12);
 		REQUIRE(testLength == sv.Length());
-		for (int i=sv.Length()-1; i>=0; i--) {
-			sv.InsertValue(i, 1, i+5);
+		for (ptrdiff_t i=sv.Length()-1; i>=0; i--) {
+			sv.InsertValue(i, 1, static_cast<int>(i+5));
 			sv.Delete(i+1);
 		}
 		for (int i=0; i<sv.Length(); i++)
